refactor: be explicit about conversion from bool

+++ dynamic_programming/memoization/can_sum.cpp
@@ -12,7 +12,7 @@ auto can_sum = [cache = std::vector<std::uint8_t>{}](
 
   return [&cache](this auto&& can_sum_, int t, const auto& v_) -> std::uint8_t {
     if (t <= 0)
-      return t == 0;
+      return static_cast<std::uint8_t>(t == 0);
 
     const auto& target_ = static_cast<std::size_t>(t);
     if (cache[target_] < 2)
